requirements.txt
# Core dependencies
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
schedule>=1.2.0

# LLM and AI
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=1.0.0
anthropic>=0.8.0

# API and UI
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0

# Utilities
python-dateutil>=2.8.0
pandas>=2.1.0
numpy>=1.24.0

# Optional: for PostgreSQL
psycopg2-binary>=2.9.0

#google api
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0